+++ docs/changelog.rst
@@ -9,13 +9,16 @@ Change Logs
     * **Bug**         -  # 107_: next_approvals api of the instance is broken
     * **Improvement** -  # 108_: Status column of transition approvals are now kept as string in the DB instead of number to maintain readability and avoid mistakenly changed ordinals.
     * **Improvement** -  # 109_: Cancel all other peer approvals that are with different branching state.
+    * **Improvement** -  # 110_: Introduce an iteration to keep track of the order of the transitions even the cycling ones. This comes with a migration that will assess the iteration of all of your existing approvals so far. According to the tests, 500 workflow objects that have 5 approvals each will take ~10 seconds.
     * **Improvement** -  # 105_: More dynamic and better way for hooks.On the fly function and hook creations, update or delete are also supported now. It also comes with useful admin interfaces for hooks and functions. This is a huge improvement for callback lovers :-)
+
 
 .. _105: https://github.com/javrasya/django-river/issues/105
 .. _106: https://github.com/javrasya/django-river/issues/106
 .. _107: https://github.com/javrasya/django-river/issues/107
 .. _108: https://github.com/javrasya/django-river/issues/108
 .. _109: https://github.com/javrasya/django-river/issues/109
+.. _110: https://github.com/javrasya/django-river/issues/110
 
 2.0.0 (Stable)
 --------------